email-sender-agent-crewai/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asttokens==3.0.0
attrs==25.3.0
auth0-python==4.10.0
babel==2.17.0
backoff==2.2.1
backrefs==5.9
bcrypt==4.3.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
chromadb==1.0.13
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
crewai==0.134.0
cryptography==45.0.4
decorator==5.2.1
distlib==0.3.9
distro==1.9.0
docstring_parser==0.16
durationpy==0.10
et_xmlfile==2.0.0
executing==2.2.0
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.5.1
ghp-import==2.1.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.73.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.1
humanfriendly==10.0
identify==2.6.12
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
instructor==1.9.0
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
json5==0.12.0
json_repair==0.47.4
jsonpickle==4.1.1
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
litellm==1.72.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.14
mkdocs-material-extensions==1.3.1
mmh3==5.1.0
mpmath==1.3.0
multidict==6.6.2
networkx==3.5
nodeenv==1.9.1
numpy==2.3.1
oauthlib==3.3.1
onnxruntime==1.22.0
openai==1.93.0
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.10.18
overrides==7.7.0
packaging==25.0
paginate==0.5.7
parso==0.8.4
pathspec==0.12.1
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
posthog==6.0.0
pre_commit==4.2.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.5
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pymdown-extensions==10.16
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyvis==0.3.2
PyYAML==6.0.2
pyyaml_env_tag==1.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.2
tomli==2.2.1
tomli_w==1.2.0
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uv==0.7.17
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.31.2
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0
